refactor(viewer): modify isSamePoints location to Annotation
