+++ libs/insight-viewer/src/Annotation/utils/getPointsUpdatedAnnotation.ts
@@ -4,7 +4,7 @@ import { getLineLength } from './getLineLength'
 import { getTextPosition } from './getTextPosition'
 import { getAnnotationPoints } from './getAnnotationPoints'
 import { getCircleCenterPoint } from './getCircleCenterPoint'
-import { isSamePoints } from '../../utils/common/isSamePoints'
+import { isSamePoints } from './isSamePoints'
 import { getMovedPoints } from './getMovedPoints'
 import { getAreaAnnotationMovedPoints } from './getAnnotationPointsAfterMoving'
 import { getCircleRadius, getCircleRadiusByMeasuringUnit } from '../utils/getCircleRadius'
