+++ libs/insight-viewer/src/Annotation/utils/isSamePoints.spec.ts
