+++ libs/insight-viewer/src/Annotation/utils/isSamePoints.ts
